break up cli commands into separate modules
