+++ yatta/cli/config.py
@@ -1,0 +1,6 @@
+import click
+
+
+@click.command()
+def config():
+    pass
