+++ yatta/cli/delete.py
@@ -1,0 +1,6 @@
+import click
+
+
+@click.command()
+def delete():
+    pass
